Migrate transform encoder/decoder tests to standalone classes
