+++ tests/unit/test_transform_encoders.py
@@ -1,0 +1,202 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2015-2016 Glencoe Software, Inc. All rights reserved.
+#
+# This software is distributed under the terms described by the LICENCE file
+# you can find at the root of the distribution bundle.
+# If the file is missing please request a copy by contacting
+# jason@glencoesoftware.com.
+#
+
+from omero_marshal import get_encoder, SCHEMA_VERSION, ROI_SCHEMA_URL
+from omero.rtypes import rdouble
+import pytest
+
+
+TRANSFORMATION_TYPE = '%s#AffineTransform' % ROI_SCHEMA_URL
+
+TRANSFORMATIONS_201501 = [
+    (
+        'matrix(1.0 0.0 0.0 1.0 0.0 0.0)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 1.0,
+            'A10': 0.0,
+            'A01': 0.0,
+            'A11': 1.0,
+            'A02': 0.0,
+            'A12': 0.0,
+        }
+    ),
+    (
+        'none',
+        None
+    ),
+    (
+        'translate(3 4)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 1.0,
+            'A10': 0.0,
+            'A01': 0.0,
+            'A11': 1.0,
+            'A02': 3.0,
+            'A12': 4.0,
+        }
+    ),
+    (
+        'translate(5)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 1.0,
+            'A10': 0.0,
+            'A01': 0.0,
+            'A11': 1.0,
+            'A02': 5.0,
+            'A12': 0.0,
+        }
+    ),
+    (
+        'scale(1.5 2.5)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 1.5,
+            'A10': 0.0,
+            'A01': 0.0,
+            'A11': 2.5,
+            'A02': 0.0,
+            'A12': 0.0,
+        }
+    ),
+    (
+        'scale(1.5)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 1.5,
+            'A10': 0.0,
+            'A01': 0.0,
+            'A11': 1.5,
+            'A02': 0.0,
+            'A12': 0.0,
+        }
+    ),
+    (
+        'rotate(45)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 0.7071067811865476,
+            'A10': 0.7071067811865475,
+            'A01': -0.7071067811865475,
+            'A11': 0.7071067811865476,
+            'A02': 0.0,
+            'A12': 0.0,
+        }
+    ),
+    (
+        'rotate(45 50 100)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 0.7071067811865476,
+            'A10': 0.7071067811865475,
+            'A01': -0.7071067811865475,
+            'A11': 0.7071067811865476,
+            'A02': 85.35533905932736,
+            'A12': -6.066017177982129,
+        }
+    ),
+    (
+        'rotate(60)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 0.5000000000000001,
+            'A10': 0.8660254037844386,
+            'A01': -0.8660254037844386,
+            'A11': 0.5000000000000001,
+            'A02': 0.0,
+            'A12': 0.0,
+        }
+    ),
+    (
+        'rotate(60 50 100)',
+        {
+            '@id': -1L,
+            '@type': TRANSFORMATION_TYPE,
+            'A00': 0.5000000000000001,
+            'A10': 0.8660254037844386,
+            'A01': -0.8660254037844386,
+            'A11': 0.5000000000000001,
+            'A02': 111.60254037844385,
+            'A12': 6.698729810778055,
+        }
+    ),
+
+]
+
+
+if SCHEMA_VERSION == "2015-01":
+    TRANSFORMATIONS = TRANSFORMATIONS_201501
+else:
+    def create_transform(a00=None, a10=None, a01=None, a11=None, a02=None,
+                         a12=None):
+        from omero.model import AffineTransformI
+        t = AffineTransformI()
+        if a00:
+            t.setA00(rdouble(a00))
+        if a10:
+            t.setA10(rdouble(a10))
+        if a01:
+            t.setA01(rdouble(a01))
+        if a11:
+            t.setA11(rdouble(a11))
+        if a02:
+            t.setA02(rdouble(a02))
+        if a12:
+            t.setA12(rdouble(a12))
+        return t
+
+    TRANSFORMATIONS = [
+        (
+            None,
+            None
+        ),
+        (
+            create_transform(),
+            None
+        ),
+        (
+            create_transform(a00=1.0, a10=2.0, a01=3.0, a11=4.0, a02=5.0,
+                             a12=6.0),
+            {
+                '@type': TRANSFORMATION_TYPE,
+                'A00': 1.0,
+                'A10': 2.0,
+                'A01': 3.0,
+                'A11': 4.0,
+                'A02': 5.0,
+                'A12': 6.0,
+                'omero:details': {'@type': 'TBD#Details'},
+            }
+        ),
+    ]
+
+
+class TestTransformEncoder():
+
+    @pytest.mark.parametrize("transform_s,transform_o", TRANSFORMATIONS)
+    def test_transforms(self, point, transform_s, transform_o):
+        point.transform = transform_s
+        encoder = get_encoder(point.__class__)
+        v = encoder.encode(point)
+        if not transform_o:
+            assert 'Transform' not in v
+        else:
+            assert v['Transform'] == transform_o
